rename the Tasks project into Tasks.Api
