+++ src/Tasks/Tasks.Configurations/ConfigurationDev.cs
@@ -3,6 +3,8 @@
     public class ConfigurationDev : ConfigurationProduction, IConfigs
     {
         public new string DB_DATABASE => "Tasks_Dev";
-        public new string URL_GUI => "http://127.0.0.1:5020";
+
+        protected new static readonly Uri _urlGui = new(@"http://127.0.0.1:5020");
+        public new Uri URL_GUI => _urlGui;
     }
 }
