+++ src/Tasks/Tasks.Configurations/IConfiguration.cs
@@ -12,9 +12,11 @@
         public string EMAIL_USERNAME { get; }
         public string EMAIL_PASSWORD { get; }
 
-        public string URL_GUI { get; }
+        public Uri URL_GUI { get; }
 
         public string REQUEST_HEADER_KEY { get; }
         public string REQUEST_HEADER_VALUE { get; }
+
+        public string IP_ADDRESS { get; }
     }
 }
